move IncomingTransactionOptions type into IncomingTransactionHelper
